backend/django_service/requirements.txt
django==5.0.14
djangorestframework==3.15.2
django-cors-headers==4.3.1
psycopg2-binary==2.9.10
redis==5.0.4
celery==5.4.0
django-celery-beat==2.7.0
django-celery-results==2.5.1
elasticsearch==8.13.2
django-auditlog==3.0.0
drf-spectacular==0.27.2
python-keycloak==3.9.1
Pillow==10.3.0
pytest==8.2.2
pytest-django==4.8.0